// web/src/lib/database.types.ts
export type Database = {
// ...
  public: {
// ...
    Enums: {
      cadre_facturation: "B2B_national" | "B2B_international" | "B2G" | "B2C"
      delai_paiement_mode: "net" | "fin_de_mois"
      devis_statut: "brouillon" | "envoyé" | "accepté" | "refusé"
      document_famille:
        | "dpgf"
        | "cctp"
        | "ppsps"
        | "doe"
        | "ccap"
        | "avenant"
        | "dgd"
      facture_statut: "brouillon" | "impayée" | "envoyée" | "payée"
      facture_statut_cycle:
        | "brouillon"
        | "deposee"
        | "recue"
        | "approuvee"
        | "refusee"
        | "paiement_transmis"
        | "encaissee"
        | "rejetee"
        | "suspendue"
      facture_type_document: "facture" | "avoir" | "acompte" | "note_frais"
      ligne_type: "ligne" | "chapitre" | "commentaire"
      logement_statut: "occupé" | "vacant" | "commune"
      metier_type: "plomberie" | "electricite" | "etancheite"
      mode_paiement:
        | "virement"
        | "cheque"
        | "especes"
        | "carte"
        | "prelevement"
        | "traite"
        | "autre"
      role_membre:
        | "admin"
        | "conducteur"
        | "technicien"
        | "lecture"
        | "secretaire"
        | "sous_traitant"
      tva_categorie: "S" | "Z" | "E" | "AE" | "K" | "G" | "O"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends (DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never) = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never